Eliminate reinstallation of .stylelintrc.json

+++ includes/install/stylelintrc.json.sh
@@ -2,23 +2,48 @@
 # Install .stylelintrc.json file
 #####################################################################################
 if [[ -e "$HOME/$STYLELINT_SETTINGS" ]]; then
-   pause_and_warn "$STYLELINT_SETTINGS file exists. Renaming to .backup$STYLELINT_SETTINGS" true
-   mv -v "$HOME/$STYLELINT_SETTINGS" "$HOME/.backup$STYLELINT_SETTINGS"
-fi
+   if [[ -e "$HOME/.backup$STYLELINT_SETTINGS" ]]; then
+      inform "$STYLELINT_SETTINGS and .backup$STYLELINT_SETTINGS exist." true
+      inform "Nothing to install. Continuining..."
+   else
+      inform "$STYLELINT_SETTINGS file exists." true
+      pause_and_warn
 
-inform "Downloading $STYLELINT_SETTINGS..."
-curl -O https://gist.githubusercontent.com/code-warrior/a766f7c32bab9a82b467601800b00a46/raw/768717143df9db9c593dabb38c3c7aa63c87f66b/"$STYLELINT_SETTINGS"
+      inform "Renaming to .backup$STYLELINT_SETTINGS..."
+      mv -v "$HOME/$STYLELINT_SETTINGS" "$HOME/.backup$STYLELINT_SETTINGS"
 
-if [[ -e "$STYLELINT_SETTINGS" ]]; then
-   inform "$STYLELINT_SETTINGS downloaded successfully."
+      inform "Downloading $STYLELINT_SETTINGS..."
+      curl -O https://gist.githubusercontent.com/code-warrior/a766f7c32bab9a82b467601800b00a46/raw/768717143df9db9c593dabb38c3c7aa63c87f66b/"$STYLELINT_SETTINGS"
 
-   inform "Installing $STYLELINT_SETTINGS..."
-   mv -v "$STYLELINT_SETTINGS" "$HOME/"
+      if [[ -e "$STYLELINT_SETTINGS" ]]; then
+         inform "$STYLELINT_SETTINGS downloaded successfully."
+
+         inform "Installing $STYLELINT_SETTINGS..."
+         mv -v "$STYLELINT_SETTINGS" "$HOME/"
+
+         if [[ -e "$STYLELINT_SETTINGS" ]]; then
+            warn "$STYLELINT_SETTINGS was not successfully installed. Please investigate, then continue."
+            pause_and_warn
+         else
+            inform "$STYLELINT_SETTINGS was installed successfully."
+         fi
+      fi
+   fi
+else
+   inform "Downloading $STYLELINT_SETTINGS..."
+   curl -O https://gist.githubusercontent.com/code-warrior/a766f7c32bab9a82b467601800b00a46/raw/768717143df9db9c593dabb38c3c7aa63c87f66b/"$STYLELINT_SETTINGS"
 
    if [[ -e "$STYLELINT_SETTINGS" ]]; then
-      warn "$STYLELINT_SETTINGS was not successfully installed. Please investigate, then continue."
-      pause_and_warn
-   else
-      inform "$STYLELINT_SETTINGS was installed successfully."
+      inform "$STYLELINT_SETTINGS downloaded successfully."
+
+      inform "Installing $STYLELINT_SETTINGS..."
+      mv -v "$STYLELINT_SETTINGS" "$HOME/"
+
+      if [[ -e "$STYLELINT_SETTINGS" ]]; then
+         warn "$STYLELINT_SETTINGS was not successfully installed. Please investigate, then continue."
+         pause_and_warn
+      else
+         inform "$STYLELINT_SETTINGS was installed successfully."
+      fi
    fi
 fi
